requirements.txt
# install at once with:
# pip install -r requirements.txt
python-telegram-bot
requests
